requirements.txt
-r kattis-cli/requirements.txt

requests
beautifulsoup4
PyYAML
